requirements.txt
python-dotenv>=1.0.0
streamlit>=1.30.0
requests>=2.31.0
pydantic>=2.5.0
# langchain>=0.1.16
# langchain-core>=0.1.48
# langchain-classic>=0.0.7
# langchain-groq>=0.1.0
rich>=13.7.0
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-groq==0.1.9
